support 64bit build

+++ lda/ldacvb0_cpp/ldacvb0_test/test.cpp
@@ -151,7 +151,7 @@ CYBOZU_TEST_AUTO(test_lda_cvb0_initialization)
 	docs.vocabularies.add("b");	// 1
 	docs.vocabularies.add("c");	// 2
 
-	int V = docs.vocabularies.size();
+	size_t V = docs.vocabularies.size();
 	CYBOZU_TEST_EQUAL(V, 3);
 
 	{
